RSFML-2 Refactor createTextList into two separate methods
Method splitIntoLines is used to ensure the string will wrap at word boundaries, based on the maxLineWidth.
Method createAlignedTextList is used to build the necessary Text objects, such that the input list of strings is treated as a single block of text.  The text block will respect the position, scale, rotation, and alignment properties.